vm: shared upvalue cells + minified-bundle fixes; React 18 now runs
The headline: closures capturing the same variable now share one
Lua-style Rc'd cell (open-upvalue registry keyed by stack slot;
transitive captures clone the parent's handle). Previously each
closure got its own copy at close time, so sibling closures
communicating through a captured variable — every scheduler and
state machine in a minified bundle — read stale values, and the old
SetUpvalue 'broadcast to every closure with a closed cell at this
index' hack corrupted unrelated captures (react-dom's captured ef
became an event timestamp). Every stack shrink now goes through
truncate_stack(), which closes dying cells first — 104 call sites,
previously ~15 of them leaked stale absolute indices.

Found by running React 18 against the Copper browser and fixed along
the way, each with a minimal repro first:
- upvalue descriptor local-slot index u8 -> u16 (>=255 locals in the
  enclosing function made closures capture the wrong variable), and
  new GetUpvalueWide/SetUpvalueWide opcodes for >255 captures per
  closure.
- Function.prototype.bind kept only the chunk index of the packed
  function value, severing the bound function from its upvalues
  (React's performSyncWorkOnRoot.bind(null, root) ran with no
  captures). The full packing now round-trips through FunctionKind.
- typeof fell back to the globals map only: it now sees globalThis
  properties (UMD wrappers' `global.React = {}`), matching GetGlobal.
- Array(n) called without `new`: the (-536..=-500) global-fn range
  swallowed sentinel -507, which doubles as Array.isArray there —
  React.createElement's children packing got a boolean. Plain Array
  calls now construct, per spec.
- Object.assign is extractable as a value (`var assign=Object.assign`
  is the standard bundle prologue); Promise.prototype carries real
  then/catch/finally values (Preact lifts then via .bind).
- 'is not a constructor' errors now carry line/pc/chunk; new debug
  aids: ZINC_DISASM_CHUNK=<name|index|*>, ZINC_WATCH_UV,
  ZINC_WATCH_SLOT (all cached env lookups, zero hot-path cost).

test262: 14285/15855 (90.1%) — identical to baseline, no regressions.
All 8 unit suites pass.

+++ src/compiler/chunk.rs
@@ -61,7 +61,8 @@ pub struct Chunk {
 pub struct UpvalueDescriptor {
     /// Index: if `is_local` is true, this is a local slot in the *enclosing* function.
     /// If false, this is an upvalue index in the enclosing function's upvalue list.
-    pub index: u8,
+    /// u16 — enclosing functions in minified bundles exceed 256 local slots.
+    pub index: u16,
     /// True if capturing directly from the enclosing function's locals.
     /// False if capturing from the enclosing function's upvalues (transitive capture).
     pub is_local: bool,
